requirements.txt
pip==23.1.2
attrs==25.3.0
pytest==8.4.1
allure-python-commons==2.15.0
Appium-Python-Client==1.3.0
selenium==3.141.0
urllib3==1.26.18
Pygments==2.19.2
pluggy==1.6.0
iniconfig==2.1.0
certifi==2025.8.3
idna==3.10
charset-normalizer==3.4.3
packaging==25.0
setuptools==65.5.0
allure-pytest==2.15.0
requests==2.32.5
webdriver-manager==4.0.2
python-dotenv==1.1.1